safari-534.53-branch/LayoutTests/svg/animations/animate-gradient-transform-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Tests if gradientTransform of a gradient is animateable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
